+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Add `DataDocument.getChildElements()` helper method that returns a list of the child elements, skipping all the comments.
+
 ### [v0.01](https://github.com/realityforge/zifnab/tree/v0.01) (2019-08-06)
 [Full Changelog](https://github.com/realityforge/zifnab/compare/b24bbdea2237c119e17341e5597c42b21b76a9c9...v0.01)